Added ability to configure Environment explicitly - #125
There is now an option to configure environment that would be used by app-gradle plugin explicitly using following syntax:
```
appengine{
  environment = ['standard'|'flexible']
}
```

If no environment was supplied fallback to the previous behaviour would be used.
If environment doesn't match the values above configuration exception would be thrown as environment is an enum.